Simplify deps
